requirements.txt
selenium==4.18.1
requests==2.31.0
pyinstaller==6.5.0
python-dotenv==1.0.1
dnspython==2.4.2

# Dependências de teste
pytest==8.0.0
pytest-mock==3.12.0 